requirements.txt
# LangChain
langchain==1.2.13
langchain-community==0.4.1
langchain-core==1.2.20
langchain-experimental==0.4.1
langchain-groq==1.1.2
langchain-huggingface==1.2.1
langchain-text-splitters==1.1.1

# LLM
groq==0.37.1

# Embeddings et Reranking
sentence-transformers==5.3.0
torch==2.10.0
transformers==5.3.0

# Vector Store
faiss-cpu==1.13.2

# Documents
pypdf==6.9.1
python-docx==1.2.0

# Interface
streamlit==1.55.0

# Utilitaires
python-dotenv==1.2.2
numpy==2.4.3